Configures lidar with angle offset as double
Updates the lidar configuration to use a double for the angle offset, allowing for more precise adjustments.

This change affects multiple modules, including core libraries, robot configurations, and simulator settings, ensuring consistency across the system.

+++ robot-system-lib-parent/robot-system-lib-bouchon/src/main/java/org/arig/robot/system/capteurs/LidarTelemeterBouchon.java
@@ -98,7 +98,7 @@ public class LidarTelemeterBouchon implements ILidarTelemeter {
   }
 
   @Override
-  public void setConfiguration(boolean reverse, int offsetAngle, int excludeLowerThan, int excludeGreaterThan) {
+  public void setConfiguration(boolean reverse, double offsetAngle, int excludeLowerThan, int excludeGreaterThan) {
     // NOOP
   }
